Refactor logging routine
The logging routine has been simplified, with three main changes

1) Using a context manager to take care of closing the file and releasing the
resources once we are done using them.

2) Using a defaultdict instead of a regular dict when counting up the
contributions of each user to eliminate the noisy "if key not in dict, key[dict]
= 1" test.

3) Using f-strings and print statements to generate the strings to write to the
file instead of just string concatenation. That takes care of getting the right
separators between the different columns of the table