Build URL templates outside the loop
We have two separate api endpoints we're looking at; one where we can
successfully walk the tree of comments, and one where we have to ask for each
comment by id one at a time. The template for these two urls is now built
outside the loop, with the comment id being added in throught string.format
inside the loop.