Use id of get rather than gz as entry point
As originally written, this script used the id of the comment after the get as
its entry point, rather than the id of the get itself. That's not ideal, since
most of the links between threads are to the get rather than to the comment
after the get.

A simple fix would have been to wrap the initialisation in a call to "getchild"
or similar, and then use the existing code to do everything else. That assumes
that every get has a comment after it (which is probably true), but it's fairly
inelegant.

The reason for the original state is that the previous code processed batches of
comments that were half-open to the right. So if X is the comment given as
initial value, the first comments would be [X-10, X), then [X-20, X-10) etc. The
half-open intervals have the advantage that the union of two overlapping
half-open intervals is another half-open interval.

The code has therefore been converted to processing comments in downwards
half-open intervals, (X-10, X]. That means we can supply the last comment in the
chain as X.

This refactoring has had the added advantage of making the logic for manual
backtracking in the case of a broken comment chain simpler. That's because we
haven't processed the "id_main" comment when we enter that branch now, and so
the odd loop-and a half logic to skip the first comment can be eliminated.

Additionally, both the author and timestamp of the get are easier to keep track
of now, since the last comment is the get, rather than the second last comment.