Hoist get statistics out of the loop
After refactoring the code so that the target comment for the routine is the
get, we no longer need to keep track of last and second last comments inside the
loop over the thread - we can just ask who made the last comment and what was
the timestamp right away. That lets us further simplify the loop logic.