Load more comments in get finding function
Sometimes the "previous" link on a thread points to e.g. the 997 comment rather
than the get. We want to scan down to find the get, but when we do that praw
sometimes returns a MoreComments instance, to indicate that we need to make a
new request to the reddit api. We should probably do that