Use pushshift to try and recover deleted accounts and comments
Sometimes people delete their accounts, their comments, and sometimes people are
shadowbanned and everything shows up as "removed by moderator". Even in these
cases, we can access the relevant submission and comment id's - there's just no
data associated with them. Pushshift has an archive of reddit submissions and
comments, so we query it to see if it has any data on the relevant id's.