Update logging script to log n threads instead of just one
The logging script now accepts an optional parameter `-n` which determines how
many threads to log, with default value 1. After each thread, the program tries
to find the previous get and then logs that thread, until the total number of
threads is reached. The log for each thread is saved in the results folder, so
that they're all kept together.