Add option to query pushshift  instead of the reddit api
Pushshift is a huge archive of reddit posts (comments + submissions) which is
more suitable for our purposes than querying the reddit api. Since it is an
archive, it doesn't need to include functionality for adding posts, comments or
votes, and it focusses just on earching for posts and returning them.

In particular, it has quite a nice endpoint at `'submission/comment_ids'` which
returns the comment ids of all comments on a submission. We use this, combined
with the ability to retrieve comments by id in bulk to pull down a complete copy
of the thread we are interested in. In theory, we can get 1000 comments at once,
but doing this gives weird errors, so 100 is used instead. Typically, there's
still a speedup of more than a factor of 2 even with this setting.

In order to proceed from here, it was necessary to add a bunch of code to
recreate post and thread objects locally, so that the thread walking code is
able to work with the offline archive. This has the benefit of making much of
the other code simpler, since the semantics of these objects are clearer than
the strings and lists which were passed around before.

The logging code has been updated to make use of this last feature, combined
with the power of pandas dataframes.