Make walk thread return comments in the correct order
Previously, walk thread returned the comments ordred from leaf to root, and all
consumers had to reverse the list before doing anything with it. That changes
now.

Additionally, walk thread also returned raw comment objects, instead of the
extracted dictionary that all the call sites wanted anyway, so that's been
changed. I'm still on the fence about that one, but we'll see