Fix a nasty bug in psaw querying
If a comment is unavailable on psaw, then psaw will return an empty generator
when we ask for it by id. Calling next on this results in a StopIteration error,
which previously crashed the entire code. Since not all the threads are on psaw,
this is a bad thing.

At the same time, it seems liek a bad idea to rename `author` and `created_utc`
when both praw return those variables. It led to code littered with `if
hasattr()` guards. Keeping the member variables, but renaming them in the
returned dictionary seems like a better idea.