Improve psaw thread navigation
This commit adds a helper function `fetch_thread` which automatically takes care
of first trying to fetch a thread using pushshift, and if that fails it falls
back on using the reddit api. Previously, we had almost identical try/catch
logic covering this case in `db.py`, `log_thread.py` and `validate_thread.py`.

Additionally, fetch_thread now checks to see how many places the thread is
broken in after fetching the comments from psaw, and if it's only broken in a
few places, it tries to patch the thread before falling back on the reddit api.

With those two changes in place, we can remove the user flag on the command line
interface specifying whether or not to use pushshift: For 99% of cases, we want
to use it, and in that case, having it as a flag is just clutter.