Add return object when updating tree
When adding a new comment to a comment tree, the newly added comment is
returned. That lets us ask sensible questions of it, like whether or not the
comment is at the root of the tree. We can use this when filling in the gaps in
`fetch_comment_tree` as an early exit condition