Make the comment tree responsible for keeping track of the root
For various purposes, it's nice to be able to ask "is this comment at the root
of the tree". It's also nice to be able to say "give me the subtree rooted at
this comment".

Previously, responsibility for keeping track of this was scattered a bit all
over the place. Now, it lies with the comment treem which is the object that has
enough information to make that kind of decision.