Remove rules logic from validation script
With our shiny new side thread objects, we don't need to repeat the rules in the
validation script - we can just import the relevant functions and classes. That
leaves us with validate.py as basically a runner script for side thread. This is
OK, since it's nice functionality to expose