Update get_history so that it can also take timestamps into account
Get history should return enough history that a rules object can always
determine whether the next count in the sequence would be valid.

That means for the wait-n threads, at least n comments are needed, since the
next commenter can't be present in the previous n.

For the slow variants, that means getting sufficiently back in time, and this
was not possible previously. With the new ability of trees to automatically get
the relevant comments, it became possible and has now been implemented.