FIX: Ignore links to the next thread in the body of a post
Sometimes people update the body of their post to include a link to the next
thread. That can give circular lookups in the child/parent relation, since we
assume that the first link in the body of a post is a link to the previous
thread.

We could therefore be in a situation where

parent(x) = y, and parent(y) = x

This is assumed to be a DAG, so that's a problem.

We fix it by only recording thread links if the post they reference to is older
than the post they are in.