Increase responsibilities of tree class
More of the tree-related code in CommentTree has been moved to the generic tree class.

Additionally, tree already keeps track of which comments have which parents. Making it
responsible for returning the traversal [leaf, root] seems sensible

The interface of the class has also been updated. The default behaviour is now
that if a tree is asked for a node, it just returns the element in its node
dictionary, instead of trying to construct a new object.

That makes the `Node` class superfluous, and the traversal methods previously
contained there have been moved to the reddit Comment class, which we want to be
aware of its tree structure.