Fix bug in `find_missing_replies`
`find_missing_replies` used to be a method on a comment instance, rather than on
the tree. When it was moved, references to `tree` and `self` were not updated as they
should have been.

Additionally, calling `replace_more` with a limit of 0 degrades performance,
since sometimes we will only get one comment as a reply, and will therefore have
to call `add_missing_replies` again in the very next step of our traverse. That
ultimately means three api calls to retrieve just one comment - not very efficient!