Deal with broken threads when walking down submissions
Sometimes reddit returns phantom comments that don't exist.

We can check for phantom comments by getting the parent of the comment we are
currently looking at, and asking for its updated replies. If the current comment
is not one of those, then its a phantom comment. If we encounter one of those
when walking down a thread we go to the parent comment, and try walking down
from there.