Add new functions for determining the total counts in a side thread
Some threads don't have a fixed length, but instead have simple rules to
determine the total number of counts, based on the chain from known submission
to latest submission, and on the previous count.

This commit adds three of those rules (one for wave, one for increasing
sequences, and a fallback for when a thread traversal is the only way to really
know how long a thread was), and modifies the side thread class so that it can
use those rules.