FIX: Enable the walking down of threads with only one comment
When we reach the leaf of a tree, we check whether the comment we have found is
real or a ghost comment. We did this by looking at the parent od the comment,
refreshing it, and seeing what its replies were. If the leaf of the tree was
also the root, that caused some problems...