By default, do not use pushshift to fetch threads
Pushshift is three days behind reddit, so almost no information is gained when
we are updating the directory page every day.

Additionally, benchmarking shows that the superfluous api calls to pushshift to
get the comments associated with a thread cost ~200s per run. That's a lot!

So, this commit flips the default behaviour, and adds a flag if you want to use
pushshift. The previous "Only use pushshift" option is removed, as it no longer
serves any purpose.