Ignore urls in markdown links when validating counting formats
For most threads, we only care about the text that's visible to the user when we
determine whether or not a given post looks like a count.

As a specific example, if a comment consists only of soem other text plus a
reddit link, e.g. "[Late
chain?](/comments/submission_id/thread_title/comment_id)", there's a significant
risk that the url will make it seem like a valid count (because reddit urls
often do contain digits), even though the comment is not a valid count.

For most threads, the correct approach is therefore to remove all the urls from
the markdown links before testing whether the comment looks like a count.

The exception is invisible numbers, where the number itself is written in the
url field, and there is no visible text.