Make `walk_down_tree` a member function of the tree class
Conceptually the idea of "Go down a tree, picking the earliest reply each time"
clearly belongs to the tree class, rather than as a function taking a tree as
input.

The only snag has been that for our purposes, we need to check that a comment is
valid before blindly walking down the tree to it.

What we've done is therefore split the logic in two. First the tree is pruned,
to remove any invalid counts and all of their children, and then the simple tree
walking algorithm is applied.

That results in conceptually much simpler code.