Consider deleted comments to be valid, but low-priority
I've changed my mind a couple of times about what I want to do with deleted
comments in a chain.

I think this is the best approach: We consider deleted comments to
be (potentially) valid counts in the chain, but if there are any other valid
non-deleted replies on the same level, we'll return those first.