Reverse direction of tree for tracking side threads
When the code was first written, there was no easy way of walking down a tree.
So, since side threads are guaranteed to be linear, the tree was reversed, and
`walk up tree` was used instead.

That's sort of a leaky abstraction, so now that we have a walk down tree method,
we can put the trees the right way round