Fix determination of root comments in tree
The root comments of a tree are those which have no parents.

So the set difference between parents and children.

For unrooted trees, we should ignore the parents which are outside our domain of
application. That is, if the parents we find are not in the set of nodes we are
considering, we should instead look at the children of the parents we find.