Add a sensible fallback option for unknown side threads
Unknown side threads were previously just assumed to be 'decimal', and given
that name as well.

Instead, they're now given the name 'default', and the logic in get side thread
has been reordered, so that the fallback case comes at the end. That lets us
provide a helpful note if the default side thread is returned