Change default validation of side threads
The heart of this commit is the change from a default counting form of `base 10`
to `permissive`.

That means that all threads in the huge dict which assumed one or the other had
to be updated.

That also means that a lot of threads which previously could just use the
default parameters now have to be explicitly set.

Conversely, some threads for which no validation is done can now be removed from
the huge list, since they just need the new default treatment.

The point is to assume less about completely unknown threads.