Add functionality to write new threads to directory file
If ever a submission is found which has no link to a previous submission, we
consider it a potential new thread.

We walk down the chain of new threads until we hit the leaf submission, and then
we compare that with the already known submissions. If the submission is truly
new, we add it to a table at the end of the directory, creating the table if it
is not already there.

When we do this, we do not call the row.update function because that would mean
fetching all the comments of the new submission twice: once when we determine
whether the new thread is long enough, and once when we actually walk down it at
the end.

In turn, that means that we need to ensure that rows can be printed even if they
haven't been updated, which requires setting a flag when the row is initialised.