Add functionality to revive archived threads
A revived thread is one which

- Is not currently present in the directory
- Is reachable through a chain of submissions from a thread on the archive page

If we encounter a revived thread, it should go in the section for "New and
revived threads" at the bottom of the directory page.

Additionally, it should be removed from the archive page.

This means that the archive page must be updated not only if there are newly
archived threads, but also if any threads have been revived