Account for singletons
Some comments have no parents and no children, and will therefore not show up in
the comment tree at all.

But they should still be included as roots of the tree

Similarly, the singletons will not be present in tree.tree, only in tree.nodes.
So if we ever want to delete them, they should only be deleted from the nodes.