Add functionality to find the deepest comment in a tree
The rule for when a new/revived thread should be added to the directory (listed
here: /comments/oldjtv/_/h5sa13s) refer to the total number of counts made on a
submission, which translates 1-1 to the depth of the relevant comment.

Since the rules for new and revived threads can be unclear, I expect more
non-counting side threads to occur, and some of them to occur before the
relevant counts in the thread. Additionally, for new and revived threads there's
no good way of pruning the tree of comments that don't look like counts.

So, we can instead just ask for the deepest comment in the tree, and use that as
our target comment.

It's highly unlikely that 50 or 20 nested non-counting comments will be made.