Refactor SubmissionTree class
Firstly, we move the SubmissionTree class into models the module where it more
naturally belongs.

Secondly, we remove the `find latest comment` member function from it and put it
back in the row class. This is because the logic is really an intersection of
things two different objects should know about:

1) Finding the latest submission in a thread
2) Finding the latest comment on a submission

The first rightly belongs to the SubmissionTree class, the next to the
CommentTree. Superficially, the logic is similar, but there are enough edge
cases that they cannot be handled just by the Tree class.

Instead of making this logic a member of either class, this commit asks each
class for the bits that it is responsible for, and then glues it together.

To give the SubmissionTree class a bit more responsibility, and to simplify the
handling of archived row, I've added an `is_archived` function, and added an
early exit condition.

Finally,  when we have a new thread, the find latest count logic should pick the
first **countlike** comment in new threads, instead of the overall first
comment.