Refactor and remove the Table class
For interacting with the directory, the code had the following data structures:

- A Row class, which did the bulk of the heavy lifting, with a row knowing how
  to update itself to find the latest submission & comment
- A Table class, which knew how to print a collection of rows, and keep track of
  a few other details

The Table was basically a fancy list, and stored very little state apart from
what was already in the row. The thing it did keep track of was

- Whether archived rows should be printed or not
- Whether the table was in the directory or in the archive page

Crucially, both of these pieces of state were only relevant for how to print the
table.

So, instead of having a table class with a list of rows as a member, tables are
represented just as a list of rows, and I've written a function which knows how
to print a table.

The two pieces of state the table kept track of can then be passed to this
function as a parameter.

This change gets rid of a lot of boilerplate (the reimplementation of all the
list stuff), and greatly simplifies the code