Move on if updating one row fails
Updating a row night fail; this is particularly true for new threads and
revivals which are generally in a bit of a flux

Concretely, there was a problem today where updating a new thread with a single
comment failed for as yet unclear reasons

If updating a single row fails, the most sensible thing is to roll back the
state of the row to the initial state, and move on to the next ones.

This is currently done by making a copy of the initial row and then subsituting
with that copy if the update fails.