Filter the list of urls when searching for the previous get
As with side threads, sometimes users link to the current submission in the body
of the post, either to highlight a specific comment, or for shits and giggles.

Luckily, submission ids are just base36 encoded running
integers in a postgres database, so by comparing the current submisison id with
the submission id of a potential link, we can see which one was made first. Just
as for the side threads, only links to posts older than the current post should
be considered when searching for the previous get.