Refactor the thread update script
The thread update script was long, and consisted of

1. A Row class which did most of the heavy lifting
2. A function to get six months of counting history
3. Some helper functions to format titles, print rows, and read in the
    directory.

After all that, there was a bit of convoluted logic in how the directory update
actually works - first the current directory and the history are loaded, then
each thread in the current directory is updated, then the code looks for new
threads, and finally for threads which had been archived but are now revived.

This commit moves the code in the first three bullet points to more logical
places. The row class itself, and most of the directly associated functions can
now be found in `thread_directory.py`.

The history getting function is moved to thread navigation, since that's the
module generally responsible for interfacing directly with reddit.

Some of the string processing stuff, like normalising titles etc has been moved
to parsing.py

After all that, update_thread_directory no longer contains functions which would
be interesting to import from elsewhere, so the "if __name__ == '__main__'"
construct is redundant.