Use tokens for authorization rather than usernames and passwords
Reddit has different types of apps, including `script` and `installed`.

- Script apps are assumed to run on machines under the control of the developer,
  are able to hold a client secret, and typically authenticate using usernames
  and passwords. They typically interact only with the developer's account, or
  with accounts directly controlled by the developer.
- Installed apps run on machines the developer doesn't control, for and by other
  users. RIF is an example of this kind. For these apps, authentication and
  refresh tokens should be used instead of passwords.

The advantage of using this flow is that anybody who tries to download the tools
here (eg through) pip doesn't need to ceate their own app, and doesn't need to
trust this program with the entirety of their account, but only with read,
wikiread and wikiedit privileges. Read and wikiread are basically free, since
anybody could do that from a web browser anyway (apart from for private
subreddits).