Use the click library for handling the cli interface
Previously, I was using the argparse library to handle the command line
interface. The structure was a single "rcounting" parser with the separate
commands attached as subparsers. This was slightly unsatisfactory as it meant
that there was a conceptual gap between the functions that did stuff and the
command line interface to them. Looking at the structure of `scripts/cli/.py`
made this clear: It contained a lot of logic about what options each of the
subcommands took, and how to make that work properly; logic which felt out of
place there.

Similarly, the functions which actually do things were strangely bare; detached
from the context of how they were intended to be invoked: to understand what was
going on, you basically had to have both cli.py and the function open.

Using click, each function can be decorated to have meaningful parameter names
and still have a command line interface.

As a bonus, implementing the full command line interface becomes an incredibly
simple exercise in composing the three subcommands.

To avoid ambiguity, the "n" parameter of the logging script has been renamed to
"n_threads", and the short option added as a synonym