Add "side thread" parameter to log side threads
Side threads are different from the main thread in two ways:

- Their get doesn't end with 000
- Finding their basecount is basically impossible, since no format for them is
  defined

This commit adds an option to log side threads instead of main. It handles the
above issues by sidestepping them:

- The format of the get is only used to make sure that the right comment is
  linked (and not e.g. the gz or the assist). Since we don't know what the get
  looks like, the right approach is to assume the linked comment is correct, and
  not do anything fancy.

- The basecount of each thread is used in two places: For ordering threads, and
   for calculating the "official" total number of comments, ignoring any
   miscounts. For the first purpose, we can use the comment submission id
   instead. For the second, there's not much to be done.

The missing information means that the only reliable way we have of identifying
a submission is its id, and hence the side thread logs are only ever written to
an sqlite database, where that information is more easily tracked.

Finally, adding support for side threads means that it's necessary to know more
initial submissions if we want to track when we've reached the start of the
submission chain.