Fix bug in fallback get finding
When logging submission threads, the program needs to know which submission and
which comment to go to after the current one. These are generally linked in
either the body of the submission, or in one of the top level comments.

Sometimes, however, only the submission is linked, and for these cases we try to
find the last comment in that submission. When we do this, we pass a submission
id rather than a submission object, but the called function was expecting the
opposite.

This commit fixes that mismatch