Make get finding more robust
When the program has to log the next thread, it looks for links in the body of
the current thread, and in the top level comments. Sometimes there are multiple
links present, with the first one being a link to the thread, and the link to
the get not appearing until later.

This commit makes the code run through all the links until it finds one
specifying both a thread and a comment, if it can.