Use sql whenever more than one submission is logged to a single file
If the user expects the output of a log to go to a single file, and also wants
to log more than one thread, the only sensible way of doing that is through sql
output rather than csv files. At least the way it is now, where csv logs are
written to a separate file for each submission.