Set comment sort at the beginning of processing
For the ordering of comments to work, they have to be sorted with the oldest
first. Generally, this is already the case, but it's nice to make sure.
Previously, the sort was modified after all the comments were fetched, which
didn't actually do anything, apart from making praw print a nice warning
message.

Instead, we'll set the comment sort as soon as we know what submission we're
interested in.