Refactor thread logging script
The logging script was a tangled mess of behaviour which varied depending on
whether or not the intended output was a csv file or a database.

Hidden inside that mess, there was a logging class which should be responsible
for knowing whether to use csv or sql to log files; how to determine which
output file to write to in the two cases, and whether or not a submission has
already been logged.

Getting to the stage where that logging class appeared took a while