Add a directory class
When I ran pylint on the code base, I got a number of warnings in the update
thread script due to basically having too long a function.

I fixed those by splitting the directory update into more or less meaningful
stages:

1. Update the known tables in the directory
2. Find new threads
3. Find revived threads
4. Use (1) and (3) to update the archive if necessary

These stages needed some communication between them - for example, new and
revived threads are added to the last table of the directory, so that needs to
be available; and the archive is updated by adding newly archived threads, and
removing the revived ones.

There was also a bit of state to keep track of, particularly whether the archive
actually needed updating or not.

To group these things more closely together, this commit adds Paragraph and
Directory classes, with the approximate logic being:

Paragraph := Text | List(Rows)
Directory := List(Paragraphs)

As a consequence of this change, some of the helper functions in
`thread_directory.py` have become class methods of either Directory or Row. In
particular, each of these classes knows how to create its own string
representation.