Avoid getting more history than necessary when updating side threads
There's an implicit assumption built in to the thread directory that **Comments
already in the directory are valid**. We can use that assumption when updating
the side thread by validating from the descendants of the comment rather than
from the comment itself. This makes a difference for how much history it's
necessary to retrieve.