🐛: Recreate a slimmed-down comment class
I was a bit too hasty in 3795621 -- the comment class does have some use cases
throughout the code:

- It can keep track of its own depth, unlike the praw comment object
- It carries around a copy of a comment tree, and thus knows how to walk up this
  tree.

These abilities are things the rest of the code relies on, so a class has been
recreated to provide them.

At the same time, the is_root helper function has been removed, since both praw
comments and the new class expose that as an attribute.