🐛: Fix archive updates
After The Big Refactor, the code path for updating the archive was untested; now
a thread has been revived and revealed an error:

The idea is that a revived thread is moved from the archive to the main
directory. This is done by deleting it from the archive dict, and adding it to
the directory Document. However, both of these are indexed by their submission
id, so if this is done after the row has been updated, then the key of the
updated row won't match the key in the archive. And that's a bummer