🐛: Find deepest comment even when there are only root comments
This is an annoying bug. The deepest node method on the Tree class is supposed
to find the deepest node in the tree. It does this by recursively finding the
depth of each leaf, and picking the greatest one.

The base case for the comparison was a depth of zero, which fails when there are
only root comments, since these also have a depth of zero.

Now that's annoying