Update rules for how archived threads are treated
It's now possible for an archived thread to be completed, and for the next
thread to start as it normally should. This commit therefor scraps the automatic
rule that if a thread has an archived direct parent, then it's a revival.

Instead, truly revived threads from the archive are detected like any other
revivals, based on the title