🐛: Remove implicit conversion between PRAW and local comment objects
This was a doozy. I discovered the error after updating the directory failed.
Looking through the logs showed that the culprit was a thread with just a single
comment. A superficial inspection of the code didn't immediately show anything
wrong, so I had to dig deeper.

When calling `walk up tree`, the old code incorrectly converted the starting
point to praw comment object. Usually, this is not a problem since it correctly
handled all other comments in the tree. However, if the starting point was ever
accessed and any of the local-comment-specific functions or attributes were
accessed, the result would be a crash.

And that's exactly what happened with the one-comment threads.