Implement functions for doing weighted k-core analysis
The k shells and the k cores are a concept from graph theory; the k cores of a
connected graph are the biggest subgraph that the induced degrees of each of
each node in the subgraph is at least k. The k shell is the k core\k + 1 core.

This definition plays badly with weighted graphs, so a different approach is
needed.

This commit implements the approach suggested by Garas et al in
https://iopscience.iop.org/article/10.1088/1367-2630/14/8/083030.

It should be noted that while the calculating the unweighted k cores and k
shells can be done in O(V) time, doing it for the weighted graphs makes the
asymptotics n log(n). That's because once  a node is deleted from the graph,
sorting the nodes according to the new degrees is trivial for unweighted graphs,
but requires a bisect for the weighted graph.