🐛: Fix failing calculation of total counts in no repeating digits
In e9e096d the calculation of powerball totals was fixed by altering the
function `permutation_order` to always assume that the permutations should be
lexicographically ordered.

Unfortunately, No Repeating digits uses the same helper function and doesn't
have the rule that the permutations should be lexicographically ordered, so that
fix introduced a bug.

This commit adds a parameter (`ordered`) to the helpher function and two
separate code paths so that both powerball and NRD work now.

It doesn't feel particularly elegant, but there is a lot of overlap between the
code in the two cases, so I can live with it for now