Fix 🐛: Avoid duplicate creating entries in directory
The directory updater works by:

- Updating all existing rows
- Checking for new and revived submissions and updating those

The new submissions are those which do not link to another submission; the
revived submissions are those which link to a submission older than six months.

If a submission is "new" but already in the directory, it might be duplicated.
To avoid this, there was a check where submissions which were already known did
not count as new or revived.

But if a "new" submission hit a get, then the latest submission was *not* added to
the known submissions, and so duplication could occur.

This commit fixes that