Prune removed comments from tree
Commit 165652e made removed comments have the lowest priorty in the depth first
expansion of the comment tree. That's still a bit unsatisfactory, since it
theoretically means that a removed comment could be returned, if, for example,
a script is run after the comment is removed but before the new chain is
started.

With removed comments a moderator has actively decided that these comments
should not be part of the chain, so completely excluding them from consideration
seems like a better approach.