Ignore DuplicateReplace errors for now an soldier on
Reddit has started sending commens in a way that praw is not expecting for
counts that occurred around 21st February 2023. For now there's no way to
recover from this, so the best approach is to abort the update of the specific
row, log a message and move on.

It seems that moving up a chain still works, so if tthis persists I might have
to start looking at that as a workaround.