Manually resolve all `MoreComments` instances`
Reddit has started returning weird `MoreComments instances` and praw isn't set
up to handle this. In particular, reddit sometimes returns both a comment and a
MoreComments object containing that comment. This breaks several praw
invariants, and so praw chokes. Additionally, for the spurious MoreComments
instances, praw is having trouble assigning them a submission instance.

The solution is to handle these things myself: loop over all replies, and every
time a `MoreComments` instance is encountered, replace it by its expansion. The
resulting flat list of comments has no tree structure, but the tree classes in
the `rcounting` tools can handle that. Additionally, they don't care about
duplicate entries in the tree; the latest one is just chosen as valid, and as
long as it refers to the correct comment, there are no issues.

Hopefully this change means that the directory updater will start working again.