Move to a single source of truth for the package version
Previously, the version was written in both `setup.cfg` and in the user agent
string given to reddit. Since every installed package has to declare a version,
there's no reason to hardcode it in the user agent when it can be loaded using
the importlib library. That makes it less likely that things will go out of sync