Add functionality to find errors in the chain
It happens semi-often that we want to check the count on a given chain, and make
sure it's in the right place. Doing that requires a method str -> int that goes
from a comment body to the number of the count in the thread sequence. Once you
have that, the errors in the chain are those where the calculated count and the
position in the submission don't match. To only find new errors, the following
two filters are also applied:

- Only look at counts which aren't the successor, S, of the previous count
- Only look at counts which aren't S(S(n-2))

That's because either following along from the previous count or just adding two
to your own previous count are very normal methods of counting, so these
shouldn't be treated as new errors.

In the same commit, for some reason, I've also decided to sort all the
dictionaries at the bottom of the file alphabetically. It's not a bad idea, but
it should probably have been its own thing.