Factor out function to identify the previous submission
When logging threads, there's a bit of logic that goes into deciding whether or
not a given submission has already been done. Doing this doesn't require access
to the get of that submission. To avoid doing unnecessary work, the "find the
previous submission" logic has been extracted to its own function for use in
other modules