Remove links from body text before logging
It's nice to be able to follow along with how far along the thread logging is.
It's even nicer not to have to look at raw markdown links when we do so.

This commit strips the link from what's printed to the terminal, so that only
the description is shown. That's a lot easier to follow