Add a way of checking whether a node is present in a tree
Sometimes we might want to ask whether or not a node is present in a tree. For
example, before asking the tree to delete the node (which will otherwise raise
an error).

That requires the tree class to implement the `__contains__` dunder, which is
what this commit does.

The reason this is relevant is that there's a hand-curated list of nodes which
aren't relevant for the thread directory. These are slowly ageing out of the
time period that's automatically collected by the updater, and thus aren't found
in the tree any more.

This commit fixes the errors I've been getting, since it lets the calling code
in `rcounting/scripts/update_thread_directory` add a test like

if node in tree:
   tree.delete(node)

The alternative approach would be to update the deletion code to silently ignore
attempted deletions of missing nodes, which doesn't seem entirely sound.